+++ src/pystra/__init__.py
@@ -56,6 +56,7 @@ from .ls import *
 from .ss import *
 from .sensitivity import *
 from .activelearning import *
+from .system import *
 
 # Calibration
 from .loadcomb import *